docs(automl,autorag): Add S3 endpoint SSRF protection documentation
Adds comprehensive security documentation for S3 endpoints covering HTTPS enforcement,
private IP blocking, and other SSRF protections. Also improves S3 repository test coverage.

+++ packages/automl/bff/internal/repositories/s3_test.go
@@ -269,7 +269,7 @@ func TestCollectBooleanValues(t *testing.T) {
 				{"0", "value"},
 			},
 			colIndex: 0,
-			expected: []interface{}{0, 1},
+			expected: []interface{}{int64(0), int64(1)},
 		},
 		{
 			name: "string booleans",
